+++ src/main/scala/io/viash/ViashTest.scala
@@ -67,10 +67,16 @@ object ViashTest extends Logging {
     verbosityLevel: Int = 6,
     parentTempPath: Option[Path] = None, 
     cpus: Option[Int], 
-    memory: Option[String]
+    memory: Option[String],
+    dryRun: Option[Boolean] = None,
+    deterministicWorkingDirectory: Option[String] = None,
   ): ManyTestOutput = {
     // create temporary directory
-    val dir = IO.makeTemp("viash_test_" + config.functionality.name, parentTempPath)
+    val dir = IO.makeTemp(
+      name = deterministicWorkingDirectory.getOrElse(s"viash_test_${config.functionality.name}_"),
+      parentTempPath = parentTempPath,
+      addRandomized = deterministicWorkingDirectory.isEmpty
+    )
     if (!quiet) infoOut(s"Running tests in temporary directory: '$dir'")
 
     // set version to temporary value
@@ -98,7 +104,8 @@ object ViashTest extends Logging {
       setupStrategy = setupStrategy.getOrElse("cachedbuild"),
       verbosityLevel = verbosityLevel,
       cpus = cpus,
-      memory = memory
+      memory = memory,
+      dryRun = dryRun
     )
     val count = results.count(_.exitValue == 0)
     val anyErrors = setupRes.exists(_.exitValue > 0) || count < results.length
@@ -147,7 +154,8 @@ object ViashTest extends Logging {
     setupStrategy: String, 
     verbosityLevel: Int, 
     cpus: Option[Int], 
-    memory: Option[String]
+    memory: Option[String],
+    dryRun: Option[Boolean]
   ): ManyTestOutput = {
     val fun = config.functionality
 
@@ -317,7 +325,10 @@ object ViashTest extends Logging {
           val subTmp = Paths.get(newDir.toString, "tmp")
           Files.createDirectories(subTmp)
 
-          val cmd = Seq(executable) ++ Seq(cpus.map("---cpus=" + _), memory.map("---memory="+_)).flatMap(a => a)
+          val cmd = if (dryRun.getOrElse(false))
+            Seq("echo", "Running dummy test script")
+          else
+            Seq(executable) ++ Seq(cpus.map("---cpus=" + _), memory.map("---memory="+_)).flatMap(a => a)
           logger(s"+${cmd.mkString(" ")}")
           val exitValue = Process(
             cmd,
@@ -340,4 +351,5 @@ object ViashTest extends Logging {
 
     ManyTestOutput(buildResult, testResults)
   }
+
 }