+++ src/main/scala/io/viash/platforms/DockerPlatform.scala
@@ -122,11 +122,11 @@ case class DockerPlatform(
       +| `ifneedbebuild` |  Build the image if it does not exist locally.
       +| `ifneedbecachedbuild` | Build the image with caching enabled if it does not exist locally, with caching enabled.
       +| `alwayspull` / `pull` / `p` |  Try to pull the container from [Docker Hub](https://hub.docker.com) or the @[specified docker registry](docker_registry).
-      +| `alwayspullelsebuild` / `pullelsebuild` |  Try to pull the image from a registry and build it if it doesn't exist.
-      +| `alwayspullelsecachedbuild` / `pullelsecachedbuild` |  Try to pull the image from a registry and build it with caching if it doesn't exist.
+      +| `alwayspullelsebuild` / `pullelsebuild` |  Try to pull the image from a registry and build it if it does not exist.
+      +| `alwayspullelsecachedbuild` / `pullelsecachedbuild` |  Try to pull the image from a registry and build it with caching if it does not exist.
       +| `ifneedbepull` |  If the image does not exist locally, pull the image.
-      +| `ifneedbepullelsebuild` |  If the image does not exist locally, pull the image. If the image does exist, build it.
-      +| `ifneedbepullelsecachedbuild` | If the image does not exist locally, pull the image. If the image does exist, build it with caching enabled.
+      +| `ifneedbepullelsebuild` |  Do nothing if the image exists locally. Else, try to pull the image from a registry. Otherwise build the image from scratch.
+      +| `ifneedbepullelsecachedbuild` | Do nothing if the image exists locally. Else, try to pull the image from a registry. Otherwise build the image with caching enabled.
       +| `push` | Push the container to [Docker Hub](https://hub.docker.com)  or the @[specified docker registry](docker_registry).
       +| `pushifnotpresent` | Push the container to [Docker Hub](https://hub.docker.com) or the @[specified docker registry](docker_registry) if the @[tag](docker_tag) does not exist yet.
       +| `donothing` / `meh` | Do not build or pull anything.
@@ -476,7 +476,7 @@ case class DockerPlatform(
     val preRun =
       if (resolve_volume == Automatic) {
         val detectMounts = args.flatMap {
-          case arg: FileArgument if arg.multiple =>
+          case arg: FileArgument if arg.multiple && arg.direction == Input =>
             // resolve arguments with multiplicity different from singular args
             val viash_temp = "VIASH_TEST_" + arg.plainName.toUpperCase()
             val chownIfOutput = if (arg.direction == Output) "\n    VIASH_CHOWN_VARS+=( \"$var\" )" else ""
@@ -611,7 +611,7 @@ case class DockerPlatform(
            |function ViashPerformChown {
            |  if (( $${#VIASH_CHOWN_VARS[@]} )); then
            |    set +e
-           |    eval docker run --entrypoint=chown $dockerArgs$volExtraParams $fullImageID "$$(id -u):$$(id -g)" --silent --recursive $${VIASH_CHOWN_VARS[@]}
+           |    eval docker run --entrypoint=bash $dockerArgs$volExtraParams $fullImageID -c "'chown $$(id -u):$$(id -g) --silent --recursive $${VIASH_CHOWN_VARS[@]}'"
            |    set -e
            |  fi
            |}